+++ tests/multiple_queue.cpp
@@ -1,13 +1,13 @@
 #include "multiple_queue.hpp"
 
-TEST_F(MultipleQueue, SingleThreadStartStop)
+TEST_F(ThreadPoolTest, SingleThreadStartStop)
 {
   ASSERT_FALSE(single_thread_pool->is_stopped());
   single_thread_pool->stop();
   ASSERT_TRUE(single_thread_pool->is_stopped());
 }
 
-TEST_F(MultipleQueue, SingleThreadSingleTask)
+TEST_F(ThreadPoolTest, SingleThreadSingleTask)
 {
   std::future<bool> result;
 
@@ -16,7 +16,7 @@ TEST_F(MultipleQueue, SingleThreadSingleTask)
   ASSERT_TRUE(result.get());
 }
 
-TEST_F(MultipleQueue, SingleThreadMultipleTask)
+TEST_F(ThreadPoolTest, SingleThreadMultipleTask)
 {
   std::size_t nb_tests = 10;
   std::vector<std::future<bool>> results;
@@ -28,14 +28,14 @@ TEST_F(MultipleQueue, SingleThreadMultipleTask)
     ASSERT_TRUE(results[i].get());
 }
 
-TEST_F(MultipleQueue, MultipleThreadStartStop)
+TEST_F(ThreadPoolTest, MultipleThreadStartStop)
 {
   ASSERT_FALSE(multiple_thread_pool->is_stopped());
   multiple_thread_pool->stop();
   ASSERT_TRUE(multiple_thread_pool->is_stopped());
 }
 
-TEST_F(MultipleQueue, MultipleThreadSingleTask)
+TEST_F(ThreadPoolTest, MultipleThreadSingleTask)
 {
   std::future<bool> result;
 
@@ -44,7 +44,7 @@ TEST_F(MultipleQueue, MultipleThreadSingleTask)
   ASSERT_TRUE(result.get());
 }
 
-TEST_F(MultipleQueue, MultipleThreadMultipleTask)
+TEST_F(ThreadPoolTest, MultipleThreadMultipleTask)
 {
   std::size_t nb_tests = 2;
   std::vector<std::future<bool>> results;